ZEN-21590: Make ZEP DB updates persistent
During 5.0.x -> 5.1.0 upgrades, the ZEP DB updates were not persisting.
Per JP, this was because they were being run within the Zope container
which does not import the ZEP mariadb.

Would be nice to add an "update" command to the zeneventserver service
to parallel the one in zope. But that leads to the circular "have to
update the service to add the command to call to update the service..."
problem.

So instead, we'll apply any outstanding updates when starting
zeneventserver.

[git cherry-picked from PR #106]